+++ spec/lite/decorator/inference_spec.rb
@@ -2,7 +2,12 @@
 
 require "spec_helper"
 
-RSpec.describe Lite::Decorator::Infered do
+RSpec.describe Lite::Decorator::Inference do
+  let(:poro) do
+    Poro
+      .include(described_class)
+      .new
+  end
   let(:user) do
     User
       .include(described_class)
@@ -11,12 +16,14 @@ RSpec.describe Lite::Decorator::Infered do
 
   describe ".decorator_class" do
     it "returns infered decorator name" do
+      expect(poro.decorator_class).to eq(PoroDecorator)
       expect(user.decorator_class).to eq(UserDecorator)
     end
   end
 
   describe ".decorator" do
     it "returns infered decorator class" do
+      expect(poro.decorator).to be_instance_of(PoroDecorator)
       expect(user.decorator).to be_instance_of(UserDecorator)
     end
   end