+++ spec/support/models/poro.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+class Poro
+
+  def first_name
+    "Foo"
+  end
+
+  def last_name
+    "Bar"
+  end
+
+end